% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw shapes}
\usage{
draw(geom = NULL)
}
\arguments{
\item{geom}{An object of class \code{sf} or \code{sfc}. The geometry is
imported to the app and can be edited. Imported geometries are shown with
black outlines while newly created geometries are shown in blue outlines.}
}
\value{
If multiple geometry types are drawn (points / lines / polygons),
returns a list containing \code{sfc} geometries with possible names
\code{Point}, \code{LineString}, and \code{Polygon}. If only a single
geometry type is drawn, returns a flat \code{sfc} object.
}
\description{
Opens a Shiny app that allows you to draw shapes; then returns the shape
as an object of class \code{\link[sf:st_sfc]{sfc}}.
}
\examples{
if (interactive()) {
  draw()
}
}
