% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{drawsf_ui}
\alias{drawsf_ui}
\title{Shiny components}
\usage{
drawsf_ui()
}
\arguments{
\item{input, output, session}{Internal arguments for shiny server functions.}
}
\value{
\code{drawsf_ui} returns the app UI, \code{drawsf_server} returns
nothing useful and is only used for its shiny functionality.
}
\description{
UI and server components of the Shiny app. Can be used to integrate into
other Shiny apps.
}
\examples{
shiny::shinyApp(drawsf_ui, drawsf_server)
}
